Moved the contribution type enum
