+++ python/src/aac/plugins/contributions/contribution_points.py
@@ -6,8 +6,7 @@ from attr import Factory, attrib, attrs, validators
 
 from aac.cli.aac_command import AacCommand
 from aac.lang.definitions.definition import Definition
-from aac.plugins.contributions.contribution_type import ContributionType
-from aac.plugins.contributions.contribution_types import PrimitiveValidationContribution, DefinitionValidationContribution
+from aac.plugins.contributions.contribution_types import ContributionType, PrimitiveValidationContribution, DefinitionValidationContribution
 from aac.plugins.contributions.plugin_contribution import PluginContribution
 
 
@@ -74,7 +73,7 @@ class ContributionPoints:
 
         Args:
             plugin_name (str): The name of the plugin that is providing the definition validation.
-            validation (list[DefinitionValidationContribution]): The definition validations that are being registered.
+            validations (list[DefinitionValidationContribution]): The definition validations that are being registered.
         """
 
         def validate(validation: Any) -> bool:
@@ -143,7 +142,7 @@ class ContributionPoints:
 
         Args:
             plugin_name (str): The name of the plugin that is providing the primitive validation.
-            validation (list[PrimitiveValidationContribution]): The definition validations that are being registered.
+            validations (list[PrimitiveValidationContribution]): The definition validations that are being registered.
         """
 
         def validate(validation: Any) -> bool:
