+++ python/src/aac/plugins/contributions/contribution_types/__init__.py
@@ -1,9 +1,11 @@
 """Module for contribution-type classes."""
 from aac.plugins.contributions.contribution_types.rule_validation import DefinitionValidationContribution
 from aac.plugins.contributions.contribution_types.type_validation import PrimitiveValidationContribution
+from aac.plugins.contributions.contribution_types.contribution_type import ContributionType
 
 
 __all__ = (
+    ContributionType.__name__,
     DefinitionValidationContribution.__name__,
     PrimitiveValidationContribution.__name__,
 )
