+++ python/tests/plugins/test_plugin_contribution.py
@@ -1,6 +1,6 @@
 from unittest.case import TestCase
 
-from aac.plugins.contributions.contribution_type import ContributionType
+from aac.plugins.contributions.contribution_types import ContributionType
 from aac.plugins.contributions.plugin_contribution import PluginContribution
 
 from tests.helpers.parsed_definitions import create_schema_definition, create_validation_definition
